feat: update neural network to account for 150 input dimension (ensuring we don't compress too quickly)
